+++ SGEyes/file_activity/watcher/folder_name_watcher.hpp
@@ -7,7 +7,7 @@ SAIGON_NAMESPACE_BEGIN
 class folder_name_watcher : public directory_watcher_base
 {
 public:
-	void set_capacity(unsigned int capacity);
+	void set_capacity(size_t capacity);
 	model_file_info2& get_add();
 	model_file_info2& get_remove();
